data_preprocessing/MAIT hackathon project/requirements.txt
pymongo[srv]
dnspython
pandas
numpy
textblob
python-dotenv
certifi
